requirements.txt
zhipuai
dashscope
openai-whisper
tqdm
pydantic
python-dotenv
openai
jupyterlab